Read stats sequentially to get around strange concurrency bug (see below)
When reading a collection of 4000 files from a USB drive, with "readStats" concurrency > 1,
it often reports the <mtime> of 1 or 2 files off by a few days. They're always different files,
and the <mtime> value is off by an exact number of days, e.g. 60 days to the millisecond.

I can't figure out the exact root cause yet, but the problem disappears if reading stats sequentially.